Refactor notification system and interfaces
- Added `UpdateRangeAsync` method to `INotificationRepository`.
- Restructured `INotificationService` to use specific services.
- Updated service registrations in `ServiceServiceCollectionExtensions`.
- Modified controllers to utilize new notification services.
- Implemented batch update functionality in `NotificationRepository`.
- Created new notification service interfaces for better organization.
- Enhanced `NotificationService` logic to leverage new services.
- Overall improvements for maintainability and separation of concerns.
