+++ SchoolMedicalServer.Abstractions/IServices/IAppointmentNotificationService.cs
@@ -1,0 +1,10 @@
+﻿using SchoolMedicalServer.Abstractions.Dtos.Notification;
+
+namespace SchoolMedicalServer.Abstractions.IServices
+{
+    public interface IAppointmentNotificationService
+    {
+        Task<NotificationResponse> SendAppointmentNotificationToNurseAsync(NotificationRequest request);
+        Task<NotificationResponse> SendAppointmentNotificationToParentAsync(NotificationRequest request);
+    }
+}
