+++ SchoolMedicalServer.Abstractions/IServices/IHealthCheckNotificationService.cs
@@ -1,0 +1,11 @@
+﻿using SchoolMedicalServer.Abstractions.Dtos.Notification;
+
+namespace SchoolMedicalServer.Abstractions.IServices
+{
+    public interface IHealthCheckNotificationService
+    {
+        Task<IEnumerable<NotificationResponse>> SendHealthCheckNotificationToParents(IEnumerable<NotificationRequest> requests);
+        Task<IEnumerable<NotificationResponse>> SendHealthCheckNotificationToNurses(IEnumerable<NotificationRequest> requests);
+        Task<NotificationResponse> SendHealthCheckResultNotificationToParent(NotificationRequest request);
+    }
+}
