+++ SchoolMedicalServer.Abstractions/IServices/IMedicalEventNotificationService.cs
@@ -1,0 +1,9 @@
+﻿using SchoolMedicalServer.Abstractions.Dtos.Notification;
+
+namespace SchoolMedicalServer.Abstractions.IServices
+{
+    public interface IMedicalEventNotificationService
+    {
+        Task<NotificationResponse> SendMedicalEventNotificationToParentAsync(NotificationRequest request);
+    }
+}
