+++ SchoolMedicalServer.Abstractions/IServices/IVaccinationNotificationService.cs
@@ -1,0 +1,11 @@
+﻿using SchoolMedicalServer.Abstractions.Dtos.Notification;
+
+namespace SchoolMedicalServer.Abstractions.IServices
+{
+    public interface IVaccinationNotificationService
+    {
+        Task<IEnumerable<NotificationResponse>> SendVaccinationNotificationToParents(IEnumerable<NotificationRequest> requests);
+        Task<IEnumerable<NotificationResponse>> SendVaccinationNotificationToNurses(IEnumerable<NotificationRequest> requests);
+        Task<NotificationResponse> SendVaccinationObservationNotificationToParent(NotificationRequest requests);
+    }
+}
