+++ SchoolMedicalServer.Api/Bootstrapping/ServiceServiceCollectionExtensions.cs
@@ -19,7 +19,6 @@ namespace SchoolMedicalServer.Api.Bootstrapping
             services.AddScoped<IMedicalRegistrationService, MedicalRegistrationService>();
             services.AddScoped<IMedicalEventService, MedicalEventService>();
             services.AddScoped<IMedicalInventoryService, MedicalInventoryService>();
-            services.AddScoped<INotificationService, NotificationService>();
             services.AddScoped<IVaccinationDetailsService, VaccinationDetailsService>();
             services.AddScoped<IVaccinationScheduleService, VaccinationScheduleService>();
             services.AddScoped<IVaccinationRoundService, VaccinationRoundService>();
@@ -31,6 +30,13 @@ namespace SchoolMedicalServer.Api.Bootstrapping
             services.AddScoped<IAdminDashboardService, AdminDashboardService>();
             services.AddScoped<IManagerDashboardService, ManagerDashboardService>();
             services.AddScoped<INurseDashboardService, NurseDashboardService>();
+            services.AddScoped<IHealthCheckNotificationService, HealthCheckNotificationService>();
+            services.AddScoped<IVaccinationNotificationService, VaccinationNotificationService>();
+            services.AddScoped<IUserNotificationService, UserNotificationService>();
+            services.AddScoped<INotificationHelperService, NotificationHelperService>();
+            services.AddScoped<IAppointmentNotificationService, AppointmentNotificationService>();
+            services.AddScoped<IMedicalEventNotificationService, MedicalEventNotificationService>();
+            services.AddScoped<IMedicalRegistrationNotificationService, MedicalRegistrationNotificationService>();
             return services;
         }
     }
