--- SchoolMedicalServer.Infrastructure/Services/NotificationService.cs
@@ -1,544 +1,0 @@
-using SchoolMedicalServer.Abstractions.Dtos.Notification;
-using SchoolMedicalServer.Abstractions.Dtos.Pagination;
-using SchoolMedicalServer.Abstractions.Entities;
-using SchoolMedicalServer.Abstractions.IRepositories;
-using SchoolMedicalServer.Abstractions.IServices;
-
-namespace SchoolMedicalServer.Infrastructure.Services
-{
-    public class NotificationService(
-        IBaseRepository baseRepository,
-        INotificationRepository notificationRepository,
-        IUserRepository userRepository,
-        IAppointmentRepository appointmentRepository,
-        IMedicalRegistrationRepository medicalRegistrationRepository,
-        IMedicalRegistrationDetailsRepository medicalRegistrationDetailsRepository,
-        IMedicalEventRepository medicalEventRepository,
-        IStudentRepository studentRepository,
-        IVaccinationResultRepository resultRepository,
-        IVaccinationRoundRepository roundRepository,
-        IVaccinationScheduleRepository scheduleRepository,
-        IVaccinationDetailsRepository detailsRepository,
-        IVaccinationObservationRepository observationRepository,
-        IHealthProfileRepository profileRepository,
-        IHealthCheckScheduleRepository healthCheckScheduleRepository,
-        IHealthCheckRoundRepository healthCheckRoundRepository,
-        IHealthCheckResultRepository healthCheckResultRepository) : INotificationService
-    {
-        public async Task<PaginationResponse<NotificationResponse>> GetUserNotificationsAsync(PaginationRequest? pagination, Guid userId)
-        {
-            var totalCount = await notificationRepository.CountByUserIdAsync(userId);
-
-            if (totalCount == 0)
-            {
-                return null!;
-            }
-
-            var skip = (pagination!.PageIndex - 1) * pagination.PageSize;
-            var notifications = await notificationRepository.GetByUserIdPagedAsync(userId, skip, pagination?.PageSize ?? 10);
-
-            var result = new List<NotificationResponse>();
-
-            foreach (var notification in notifications)
-            {
-                var request = new NotificationRequest
-                {
-                    SenderId = notification.SenderId,
-                    ReceiverId = notification.UserId
-                };
-                var notiInfo = NotificationInformation(notification);
-                var sender = await SenderInformation(request)!;
-                var receiver = await ReceiverInformation(request)!;
-
-                result.Add(GetResponse(notiInfo, sender, receiver));
-            }
-
-            return new PaginationResponse<NotificationResponse>(
-                   pagination!.PageIndex,
-                   pagination.PageSize,
-                   totalCount,
-                   result
-             );
-        }
-        public async Task<NotificationResponse> SendAppoimentNotificationToNurseAsync(NotificationRequest request)
-        {
-            var appointment = await appointmentRepository.GetByIdWithStudentAsync(request.NotificationTypeId);
-            if (appointment == null)
-            {
-                return null!;
-            }
-
-            var receiver = await ReceiverInformation(request);
-            var sender = await SenderInformation(request);
-
-            var notification = new Notification
-            {
-                NotificationId = Guid.NewGuid(),
-                UserId = receiver!.UserId,
-                SenderId = sender!.UserId,
-                Title = "New Appointment Notification",
-                Content = $"You have a new appointment scheduled with Parent of {appointment.Student?.FullName} on {appointment.AppointmentDate?.ToString("d")} from {appointment.AppointmentStartTime?.ToString()} to {appointment.AppointmentEndTime?.ToString()}.",
-                SendDate = DateTime.UtcNow,
-                IsRead = false,
-                Type = NotificationTypes.Appointment,
-                SourceId = appointment.AppointmentId
-            };
-            await notificationRepository.AddAsync(notification);
-            await baseRepository.SaveChangesAsync();
-
-            var notiInfo = NotificationInformation(notification);
-            return GetResponse(notiInfo, sender, receiver);
-        }
-        public async Task<NotificationResponse> SendAppoimentNotificationToParentAsync(NotificationRequest request)
-        {
-            var appointment = await appointmentRepository.GetByIdWithStudentAsync(request.NotificationTypeId);
-
-            if (appointment == null)
-                return null!;
-
-            var receiver = await ReceiverInformation(request);
-            var sender = await SenderInformation(request);
-
-            var notification = new Notification
-            {
-                NotificationId = Guid.NewGuid(),
-                SenderId = sender!.UserId,
-                UserId = receiver!.UserId,
-                Title = appointment.CompletionStatus == false ? "Appointment Confirmation" : "Appointment Completion",
-                Content = appointment.CompletionStatus == true
-                            ? $"Your appointment with {sender!.UserName} is completed for {appointment.AppointmentDate?.ToString("d")} from {appointment.AppointmentStartTime?.ToString()} to {appointment.AppointmentEndTime?.ToString()}."
-                            : $"Your appointment with {sender!.UserName} is confirmed for {appointment.AppointmentDate?.ToString("d")} from {appointment.AppointmentStartTime?.ToString()} to {appointment.AppointmentEndTime?.ToString()}.",
-                SendDate = DateTime.UtcNow,
-                IsRead = false,
-                Type = NotificationTypes.Appointment,
-                SourceId = appointment.AppointmentId,
-            };
-            await notificationRepository.AddAsync(notification);
-            await baseRepository.SaveChangesAsync();
-
-
-            var notiInfo = NotificationInformation(notification);
-            return GetResponse(notiInfo, sender, receiver);
-        }
-        public async Task<NotificationResponse> SendMedicalRegistrationNotificationToNurseAsync(NotificationRequest request)
-        {
-            var medicalRegistration = await medicalRegistrationRepository.GetByIdAsync(request.NotificationTypeId);
-            if (medicalRegistration == null)
-            {
-                return null!;
-            }
-            var receiver = await ReceiverInformation(request);
-            var sender = await SenderInformation(request);
-            var student = await studentRepository.GetStudentInfoAsync(medicalRegistration.StudentId);
-            var notification = new Notification
-            {
-                NotificationId = Guid.NewGuid(),
-                UserId = receiver!.UserId,
-                SenderId = sender!.UserId,
-                Title = "New Medical Registration Notification",
-                Content = $"A new medical registration for {student!.StudentFullName} has been created by {sender.UserName} with medication: {medicalRegistration.MedicationName}. Please review it.",
-                SendDate = DateTime.UtcNow,
-                IsRead = false,
-                Type = NotificationTypes.MedicalRegistration,
-                SourceId = medicalRegistration.RegistrationId
-            };
-            await notificationRepository.AddAsync(notification);
-            await baseRepository.SaveChangesAsync();
-            var notiInfo = NotificationInformation(notification);
-            return GetResponse(notiInfo, sender, receiver);
-        }
-        public async Task<NotificationResponse> SendMedicalRegistrationApprovedNotificationToParentAsync(NotificationRequest request)
-        {
-            var medicalRegistration = await medicalRegistrationRepository.GetApprovedByIdWithStudentAsync(request.NotificationTypeId);
-
-            if (medicalRegistration == null)
-            {
-                return null!;
-            }
-
-            var receiver = await ReceiverInformation(request);
-            var sender = await SenderInformation(request);
-
-            var notification = new Notification
-            {
-                NotificationId = Guid.NewGuid(),
-                UserId = receiver!.UserId,
-                SenderId = sender!.UserId,
-                Title = "Medical Registration Approved",
-                Content = $"Your child's medication registration ({medicalRegistration.MedicationName}) has been approved by the nurse.",
-                SendDate = DateTime.UtcNow,
-                IsRead = false,
-                Type = NotificationTypes.MedicalRegistration,
-                SourceId = medicalRegistration.RegistrationId
-            };
-            await notificationRepository.AddAsync(notification);
-            await baseRepository.SaveChangesAsync();
-
-            var notiInfo = NotificationInformation(notification);
-            return GetResponse(notiInfo, sender, receiver);
-        }
-        public async Task<NotificationResponse> SendMedicalRegistrationCompletedNotificationToParentAsync(NotificationRequest request)
-        {
-            var medicalRegistrationDetail = await medicalRegistrationDetailsRepository.GetByIdWithRegistrationAndStudentAsync(request.NotificationTypeId);
-
-            if (medicalRegistrationDetail == null)
-            {
-                return null!;
-            }
-
-            var receiver = await ReceiverInformation(request);
-            var sender = await SenderInformation(request);
-
-            var medicationName = medicalRegistrationDetail.MedicalRegistration?.MedicationName ?? "The medication";
-            var doseNumber = !string.IsNullOrEmpty(medicalRegistrationDetail.DoseNumber) ? $" (Dose {medicalRegistrationDetail.DoseNumber})" : "";
-            var doseTime = !string.IsNullOrEmpty(medicalRegistrationDetail.DoseTime) ? $" at {medicalRegistrationDetail.DoseTime}" : "";
-            var studentName = medicalRegistrationDetail.MedicalRegistration?.Student?.FullName ?? "your child";
-
-            var notification = new Notification
-            {
-                NotificationId = Guid.NewGuid(),
-                UserId = receiver!.UserId,
-                SenderId = sender!.UserId,
-                Title = "Medication Dose Completed",
-                Content = $"A dose{doseNumber} of medication ({medicationName}) for {studentName}{doseTime} has been completed by the nurse.",
-                SendDate = DateTime.UtcNow,
-                IsRead = false,
-                Type = NotificationTypes.MedicalRegistration,
-                SourceId = medicalRegistrationDetail.MedicalRegistrationDetailsId
-            };
-            await notificationRepository.AddAsync(notification);
-            await baseRepository.SaveChangesAsync();
-
-            var notiInfo = NotificationInformation(notification);
-            return GetResponse(notiInfo, sender, receiver);
-        }
-        private NotificationResponseDto NotificationInformation(Notification notification)
-        {
-            return new NotificationResponseDto
-            {
-                NotificationId = notification.NotificationId,
-                Type = notification.Type,
-                SourceId = notification.SourceId,
-                Title = notification.Title,
-                Content = notification.Content,
-                SendDate = notification.SendDate
-            };
-        }
-        private async Task<SenderInformationResponseDto> SenderInformation(NotificationRequest request)
-        {
-            var user = await userRepository.GetByIdAsync(request.SenderId);
-            var sender = new SenderInformationResponseDto
-            {
-                UserId = user?.UserId,
-                UserName = user?.FullName
-            };
-            return sender;
-        }
-        private async Task<ReceiverInformationResponseDto> ReceiverInformation(NotificationRequest request)
-        {
-            var user = await userRepository.GetByIdAsync(request.ReceiverId);
-            var receiver = new ReceiverInformationResponseDto
-            {
-                UserId = user?.UserId,
-                UserName = user?.FullName
-            };
-            return receiver;
-        }
-        private NotificationResponse GetResponse(NotificationResponseDto noti, SenderInformationResponseDto sender, ReceiverInformationResponseDto receiver)
-        {
-            return new NotificationResponse
-            {
-                NotificationResponseDto = noti,
-                SenderInformationDto = sender!,
-                ReceiverInformationDto = receiver!
-            };
-        }
-        public async Task<NotificationResponse> SendMedicalEventNotificationToParentAsync(NotificationRequest request)
-        {
-            var medicalEvent = await medicalEventRepository.GetByIdWithStudentAsync(request.NotificationTypeId);
-            if (medicalEvent == null)
-            {
-                return null!;
-            }
-            var receiver = await ReceiverInformation(request);
-            var sender = await SenderInformation(request);
-            var notification = new Notification
-            {
-                NotificationId = Guid.NewGuid(),
-                UserId = receiver!.UserId,
-                SenderId = sender!.UserId,
-                Title = "Medical Event Notification",
-                Content = $"A medical event has been recorded for {medicalEvent.Student?.FullName} on {medicalEvent.EventDate?.ToString("d")}.",
-                SendDate = DateTime.UtcNow,
-                IsRead = false,
-                Type = NotificationTypes.MedicalEvent,
-                SourceId = medicalEvent.EventId
-            };
-            await notificationRepository.AddAsync(notification);
-            await baseRepository.SaveChangesAsync();
-
-            var notiInfo = NotificationInformation(notification);
-            return GetResponse(notiInfo, sender, receiver);
-        }
-        public async Task<bool> ReadAllNotificationsAsync(Guid userId)
-        {
-            var notifications = await notificationRepository.GetUnreadByUserIdAsync(userId);
-            if (notifications.Count == 0)
-            {
-                return false;
-            }
-            foreach (var notification in notifications)
-            {
-                notification.IsRead = true;
-                notification.ReadDate = DateTime.UtcNow;
-            }
-            await baseRepository.SaveChangesAsync();
-            return true;
-        }
-
-        public async Task<int> GetUserUnReadNotificationsAsync(Guid? userId)
-        {
-            var unreadNotis = await notificationRepository.CountUnreadByUserIdAsync(userId ?? Guid.Empty);
-            return unreadNotis;
-        }
-
-        public async Task<NotificationResponse> GetUserNotificationDetailsAsync(Guid notificationId)
-        {
-            var noti = await notificationRepository.GetByIdAsync(notificationId);
-            if (noti == null)
-            {
-                return null!;
-            }
-            var request = new NotificationRequest
-            {
-                SenderId = noti.SenderId,
-                ReceiverId = noti.UserId
-            };
-            var notiInfo = NotificationInformation(noti);
-            var sender = await SenderInformation(request);
-            var receiver = await ReceiverInformation(request);
-            return GetResponse(notiInfo, sender, receiver);
-        }
-
-        public async Task<IEnumerable<NotificationResponse>> SendVaccinationNotificationToParents(IEnumerable<NotificationRequest> requests)
-        {
-            List<NotificationResponse> responses = [];
-            foreach (var request in requests)
-            {
-                try
-                {
-                    var result = await resultRepository.GetByIdAsync(request.NotificationTypeId);
-                    var round = await roundRepository.GetVaccinationRoundByIdAsync(result!.RoundId);
-                    var schedule = await scheduleRepository.GetVaccinationScheduleByIdAsync(round!.ScheduleId);
-                    var sender = await SenderInformation(request);
-                    var receiver = await ReceiverInformation(request);
-                    var student = await studentRepository.GetStudentByHealthProfileId(result.HealthProfileId);
-                    var vaccine = await detailsRepository.GetByIdAsync(schedule!.VaccineId);
-
-                    var notification = new Notification
-                    {
-                        NotificationId = Guid.NewGuid(),
-                        UserId = receiver.UserId,
-                        SenderId = sender.UserId,
-                        Title = schedule!.Title,
-                        Content = $"Your child {student!.FullName} has received the vaccination: {vaccine!.VaccineName} on {round.StartTime?.ToString("d")}.",
-                        SendDate = DateTime.UtcNow,
-                        IsRead = false,
-                        Type = NotificationTypes.Vaccination,
-                        SourceId = result.VaccinationResultId
-                    };
-                    await notificationRepository.AddAsync(notification);
-                    await baseRepository.SaveChangesAsync();
-                    var notiInfo = NotificationInformation(notification);
-                    responses.Add(GetResponse(notiInfo, sender, receiver));
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine(ex);
-                }
-            }
-            if (responses.Count == 0)
-            {
-                return null!;
-            }
-            return responses;
-        }
-
-        public async Task<IEnumerable<NotificationResponse>> SendVaccinationNotificationToNurses(IEnumerable<NotificationRequest> requests)
-        {
-            List<NotificationResponse> responses = [];
-            foreach (var request in requests)
-            {
-                try
-                {
-                    var schedule = await scheduleRepository.GetVaccinationScheduleByIdAsync(request.NotificationTypeId);
-                    var sender = await SenderInformation(request);
-                    var receiver = await ReceiverInformation(request);
-                    if (schedule == null || sender == null || receiver == null)
-                    {
-                        continue;
-                    }
-                    var nurseRounds = schedule.Rounds.ToList().Where(r => r.NurseId == request.ReceiverId).ToList();
-                    var notification = new Notification
-                    {
-                        NotificationId = Guid.NewGuid(),
-                        UserId = receiver.UserId,
-                        SenderId = sender.UserId,
-                        Title = schedule.Title,
-                        Content = $"You have been assigned to the vaccination schedule: {schedule.Vaccine!.VaccineName} with {nurseRounds.Count} rounds.",
-                        SendDate = DateTime.UtcNow,
-                        IsRead = false,
-                        Type = NotificationTypes.Vaccination,
-                        SourceId = schedule.ScheduleId
-                    };
-                    await notificationRepository.AddAsync(notification);
-                    await baseRepository.SaveChangesAsync();
-                    var notiInfo = NotificationInformation(notification);
-                    responses.Add(GetResponse(notiInfo, sender, receiver));
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine(ex);
-                }
-            }
-            if (responses.Count == 0)
-            {
-                return null!;
-            }
-            return responses;
-        }
-
-        public async Task<NotificationResponse> SendVaccinationObservationNotificationToParent(NotificationRequest requests)
-        {
-            var observation = await observationRepository.GetVaccinationObservationByIdAsync(requests.NotificationTypeId);
-            if (observation == null)
-            {
-                return null!;
-            }
-            var result = await resultRepository.GetByIdAsync(observation.VaccinationResultId);
-            if (result == null)
-            {
-                return null!;
-            }
-            var healthProfile = await profileRepository.GetHealthProfileById(result.HealthProfileId);
-            var student = await studentRepository.GetStudentByHealthProfileId(healthProfile!.HealthProfileId);
-            var receiver = await ReceiverInformation(requests);
-            var sender = await SenderInformation(requests);
-            var notification = new Notification
-            {
-                NotificationId = Guid.NewGuid(),
-                UserId = receiver!.UserId,
-                SenderId = sender!.UserId,
-                Title = "Vaccination Observation Notification",
-                Content = $"A post-vaccination reaction has been observed for student {student!.FullName}. " +
-                      $"The reaction began at {observation.ReactionStartTime?.ToString("g") ?? "an unknown time"}, " +
-                      $"identified as {observation.ReactionType ?? "unspecified"} with a severity level of {observation.SeverityLevel ?? "unspecified"}. " +
-                      $"Intervention provided: {observation.Intervention ?? "none"}",
-                SendDate = DateTime.UtcNow,
-                IsRead = false,
-                Type = NotificationTypes.VaccinationObservation,
-                SourceId = observation.VaccinationObservationId
-            };
-            await notificationRepository.AddAsync(notification);
-            await baseRepository.SaveChangesAsync();
-            var notiInfo = NotificationInformation(notification);
-            return GetResponse(notiInfo, sender, receiver);
-        }
-
-        public async Task<IEnumerable<NotificationResponse>> SendHealthCheckNotificationToParents(IEnumerable<NotificationRequest> requests)
-        {
-            List<NotificationResponse> responses = [];
-            foreach (var request in requests)
-            {
-                var result = await healthCheckResultRepository.GetByIdAsync(request.NotificationTypeId);
-                var round = await healthCheckRoundRepository.GetHealthCheckRoundByIdAsync(result!.RoundId);
-                var schedule = await healthCheckScheduleRepository.GetHealthCheckScheduleByIdAsync(round!.ScheduleId);
-                var student = await studentRepository.GetStudentByHealthProfileId(result.HealthProfileId);
-
-                var sender = await SenderInformation(request);
-                var receiver = await ReceiverInformation(request);
-                if (schedule == null || sender == null || receiver == null)
-                {
-                    continue;
-                }
-                var notification = new Notification
-                {
-                    NotificationId = Guid.NewGuid(),
-                    UserId = receiver.UserId,
-                    SenderId = sender.UserId,
-                    Title = schedule.Title,
-                    Content = $"Your child {student!.FullName} has received the health check: {schedule!.HealthCheckType} on {round.StartTime?.ToString("d")}.",
-                    SendDate = DateTime.UtcNow,
-                    IsRead = false,
-                    Type = NotificationTypes.HealthCheckUp,
-                    SourceId = result.ResultId
-                };
-                await notificationRepository.AddAsync(notification);
-                await baseRepository.SaveChangesAsync();
-                var notiInfo = NotificationInformation(notification);
-                responses.Add(GetResponse(notiInfo, sender, receiver));
-            }
-            return responses;
-        }
-
-        public async Task<IEnumerable<NotificationResponse>> SendHealthCheckNotificationToNurses(IEnumerable<NotificationRequest> requests)
-        {
-            List<NotificationResponse> responses = [];
-            foreach (var request in requests)
-            {
-                var schedule = await healthCheckScheduleRepository.GetHealthCheckScheduleByIdAsync(request.NotificationTypeId);
-                var sender = await SenderInformation(request);
-                var receiver = await ReceiverInformation(request);
-                if (schedule == null || sender == null || receiver == null)
-                {
-                    continue;
-                }
-                var nurseRounds = schedule.Rounds.ToList().Where(r => r.NurseId == request.ReceiverId).ToList();
-                var notification = new Notification
-                {
-                    NotificationId = Guid.NewGuid(),
-                    UserId = receiver.UserId,
-                    SenderId = sender.UserId,
-                    Title = schedule.Title,
-                    Content = $"You have been assigned to the health check schedule: {schedule.HealthCheckType} with {nurseRounds.Count} rounds.",
-                    SendDate = DateTime.UtcNow,
-                    IsRead = false,
-                    Type = NotificationTypes.HealthCheckUp,
-                    SourceId = schedule.ScheduleId
-                };
-                await notificationRepository.AddAsync(notification);
-                await baseRepository.SaveChangesAsync();
-                var notiInfo = NotificationInformation(notification);
-                responses.Add(GetResponse(notiInfo, sender, receiver));
-            }
-            return responses;
-        }
-
-        public async Task<NotificationResponse> SendHealthCheckResultNotificationToParent(NotificationRequest request)
-        {
-            var result = await healthCheckResultRepository.GetByIdAsync(request.NotificationTypeId);
-            if (result == null)
-            {
-                return null!;
-            }
-            var notification = new Notification
-            {
-                NotificationId = Guid.NewGuid(),
-                UserId = request.ReceiverId,
-                SenderId = request.SenderId,
-                Title = "Health Check Result Notification",
-                Content = $"The health check result for your child {result.HealthProfile!.Student!.FullName} is now available.",
-                SendDate = DateTime.UtcNow,
-                IsRead = false,
-                Type = NotificationTypes.HealthCheckUp,
-                SourceId = result.ResultId
-            };
-            await notificationRepository.AddAsync(notification);
-            await baseRepository.SaveChangesAsync();
-            var notiInfo = NotificationInformation(notification);
-            var sender = await SenderInformation(request);
-            var receiver = await ReceiverInformation(request);
-            return GetResponse(notiInfo, sender, receiver);
-        }
-    }
-}
